Do not consider \ an escape character when in raw strings
This implementation generally considers *all* \ in \` sequences as
non-escaping and doesn't touch other occurences of \.

This has two implications:

1) \` inside other structures, such as quoted strings and comments is
also not an escape sequence

2) other escape sequences are still considered escape sequences in raw
strings

The assumption here is that neither is a problem, because other escape
sequences in raw strings do not affect syntax parsing or
fontification, and \` is not a valid escape sequence in other
structures, so should not affect parsing or fontification, either.

If either of those two assumptions turns out to be false, we'll need
to refine this implementation.