match identifiers, not type names, when highlighting struct literal
In code like `foo(&T{})`, we want to highlight T. We only want to
highlight struct literals, which can only be T or &T. Commit
3fb128afb93780e1b768b552137c950074e78129 changed this fontification rule
to match type names instead of identifiers. The difference is, that type
names may be prefixed with an asterisk or opening parenthesis, which
would lead to matching the opening parenthesis in the function call.
Revert this change.