Don't set old buffer as active after godef--find-file-line-column
godef--find-file-line-column uses find-file to open and switch to a
file, which means the new buffer becomes the current buffer.
Nevertheless, we wrapped it in with-current-buffer, which
would (erroneously) reset the current buffer to the previous one. Only
after all surrounding functions return would current-buffer be correct
again.

This behaviour was confusing and served no purpose, so stop doing it.

Closes gh-85