Add goto functions for working with function signatures
Add several functions:
- `go-goto-arguments`
- `go-goto-docstring`
- `go-goto-function`
- `go-goto-function-name`
- `go-goto-return-value`
- `go-goto-type-signature`

They are all about going to different parts of function signatures, with
their names being self descriptive. They all operate on the current
function, the definition of which is "the closest `func` statement above
point". If no `func` is found above point, the functions will jump to
the first `func` in the file. Lastly, if the point is in the docstring
of a function, the goto functions will operate on that function and not
the function above.

Additionally, some of the functions will create their targets if they are
missing:
- `go-goto-docstring` will create a docstring if it does not
  exist as well as update the docstring name for a function that has been
  renamed.
- `go-goto-return-value` will add space for adding return values.
- `go-goto-type-signature` will add spaces and a pair of parenthesis for
  adding a signature.

It should also be noted that the functions have been tested with
anonymous functions as well.

Add keymap at `C-c C-g` for goto functions.
Add mapping for the already existing `go-goto-imports` at `C-c C-g i`.
Add helper function `go--get-function-name`, which grabs the string of
the function above point.

Update README to reflect the addition of these functions.