Don't throw errors in some go-goto-* functions
By default, don't throw errors in go-goto-method-receiver and
go-goto-docstring when called interactively on anonymous functions.
Instead, assume that the user meant to run it on the surrounding named
function.

If they are called non-interactively with a prefix argument, we should
not throw an error about not being able to run on anonymous functions
either since we have explicitly asked to ignore them with the prefix
argument.