Add go-packages-function and go-packages-go-list
Allow choosing between different ways of finding installed packages.
go-packages will call go-packages-function.

The old implementation of go-packages, which looks for .a files, now
lives in go-packages-native and is the default value for
go-packages-function.

A new implementation based on calling 'go list all' was added as
go-packages-go-list. This function produces better results, but can be
significantly slower.

Closes gh-120