Support goimports' new -srcdir flag
goimports, in order to support vendoring, needs to know where the source
file resides. Since we write the buffer's contents to a temporary file,
we need to tell goimports where the file came from.

We add a new customizable variable called gofmt-is-goimports. WHen it is
set to true, we use goimports' -srcdir flag.