Add GOPATH detection
Add functions for detecting various build tools/package managers: wgo,
gb, Godep and plain GOPATH.

The new command go-set-project uses the result of the detection to set
the GOPATH environment variable. This function can be used manually, or
invoked from something like projectile's project switch hook.

Initially, the idea was to set process-environment buffer-locally, but
that's not a useful thing to do, as most elisp code first creates and
switches to a new buffer before running a command. This new buffer would
not inherit our modified process environment.

A second idea was to use hooks to change the global GOPATH
appropriately, for example when switching buffers. However, this would
probably incur performance issues, as buffers are changed a lot
internally, not just due to user interaction.

Closes gh-118