Make godoc command and completion configurable
Introduce godoc-command that allows choosing between godoc and godoc.
Introduce godoc-use-completing-read that allows turning on or off the
use of completing-read for import paths in godoc.

Default to go doc and no completion, which differs from the old enforced
combination of godoc and completion. This is done because go doc is
generally the more useful tool.

Closes gh-101