Fix call expr indenting after dangling operator.
We were indenting like

	f :=
		print(1,
		2,
	)

because the "(" follows a line with a dangling operator, so we weren't
adding an extra indent. I think this logic was to avoid giving an
extra indent when a block opening "{" follows a line with a dangling
operator. I added some extra logic to always indent the line after an
opening "(".

Similarly we were indenting like

	f :=
		int64(4 *
		3 *
		1)

because "3"'s previous line was already indented due to a dangling
operator. I added more logic to always indent in this case when the
previous line opens a "(".

Note that there are still other broken cases involving "{" because we
can't currently differentiate between a block opening curly and a
composite literal curly:

	f :=
		map[int]bool{
		1: 3,
	}

Fixes #13