Speed up go-previous-line-has-dangling-op-p
Don't call line-number-at-pos since that starts at the beginning of
the buffer and counts lines one at a time. Instead we now cache
danglingness by the position of the start of the line instead of the
line number. We might not need the cache anymore (didn't check).

We also store and return the position of the line with the dangling
operator as the hash value (will use in later commit).