Handle indentation for nested dangling operators
The dangling indent level increases for nested dangling expressions:

foo ||
  (bar ||
    baz)

Previously we only allowed a single extra indent from dangling
operators. Now when indenting the line after a dangling operator we
use the newly added go--dangling-line-opens-indent-p to determine if
the dangling line opened a new indent level.

foo ||
  (bar ||
  baz)

is now indented properly as

foo ||
  (bar ||
    baz)

When indenting the statement following the final dangling line we go
back to the first non-dangling line to determine the indent. For
example, in cases like:

if foo ||
     (bar ||
       baz) {
  // now we indent here
    // instead of here
}

Closes: #279 [via git-merge-pr]