Support trailing commas as dangling operators.
Sometimes a trailing comma behaves like a dangling
operator (i.e. indents following line). For example:

// This comma is a dangling operator since it indents the "2".
return 1,
  2

// This comma isn't a dangling operator since it doesn't indent
// the "2" (the "2" is indented due to the paren block, not the
// dangling comma).
foo(
  1,
  2,
)

// Similarly, not a dangling operator because the indent is due
// to the curlies.
Foo{
  A: 1,
  B: 2,
}

We now have special logic to sometimes treat a trailing comma as a
dangling operator. Specifically, if we are inside parens, a composite
literal or a struct definition, do not treat the comma as a dangling
operator.

Fixes #42