Update default for godoc-and-godef-command
Starting with Go 1.12, the `godoc` command no longer has a
command-line interface (its only a webserver). That
broke (godoc-and-godef) by default.

Originally `godoc` and `go doc` had non-overlapping features which is
why we kept this as `godoc` when we switched 'godoc-command over to
`go doc`. Since that time it looks like the things we need are now
supported in `go doc`, namely you can give `go doc` a filesystem path
instead of a package name.

Fixes #262