indent: handle comments after multiline case lists
Add some utility methods for dealing with lines while ignoring
trailing comments and multiline strings. In particular, use
go--line-suffix-p to check for trailing commas and colons in
go--in-case-clause-list-p so we don't get tripped up by trailing
comments.

There is a similar existing function go--backward-irrelevant, but that
skips case statements, so I didn't want to mess with it for now.