indent: fix performance in giant comments
The recently added "case" comment logic was performing poorly on giant
comments (e.g. golang/src/cmd/go/alldocs.go). It was spending a long
time searching forwards and backwards through all the comments to see
if it was attached to a "case" statement. Note that this issue only
started happening after the previous commit. Before, we were hitting
an elisp max depth exceeded error due to the recursive nature of
go--backward-irrelevant. go-previous-line-has-dangling-op-p no longer
calls go--backward-irrelevant so the performance issue became
noticeable. I made two tweaks to speed this up:

1. Assume we aren't above a case statement if comment's current indent
   is 0. This completely avoids all the work for giant doc.go style
   comments. This will mess things up a little in cases like:

{
  switch foo {
// hi
  case 123:
  }
}

where the "hi" comment (with current indent of 0) will first get
indented beyond the case statement (but if you indent again, it will
get aligned with the case). I imagine this to be very uncommon, so a
worthwhile tradeoff.

2. Make go-in-comment-p the final "cond" clause. This is a relatively
   expensive check, so this will speed up other cases of big comments
   not covered by 1.