indent: fix various cases with dangling operators
There were many uncommon cases of multiline expressions that we
weren't indenting correctly. In general, sub-expressions always get
indented after dangling operators. Sub-expressions can be explicit
with parens, or implicit if they use operators that bind tighter than
the dangling operator. And, similarly, when indented sub-expressions
end, the indent must be reduced for subsequent dangling lines.

Some examples that are now fixed:

good:         bad:

1 |           1 |
  1&            1&
    1           1

good:          bad:

(1 |           (1 |
  (2 |           (2 |
    3)) |          3)) |
  5                5

Cases like these are uncommon, but do occur in the standard library.

The fix involved a significant rewrite of go-indentation-at-point. The
main conceptual additions are logic to handle closing parens better,
and logic to indent tightly binding operators on a continuation
line. Performance was not a goal, but things seem to be faster with
the changes.

Fixes #234