indent: fix indent at top level of file
We were always adding the (current-indentation) to the computed
indent, but that is not correct for statements at the top
level (i.e. outside any function). It was causing those statements to
maintain their current indent instead of indenting to the beginning of
the line.

I added a new type of indent test so we can differentiate between the
"maintains current indent" and "indent gets corrected" cases.