Fix infinite loop in fill-region
Calling fill-region on code (as opposed to comments) was looping
forever. This is because fill-region keeps going until it reaches the
end of the region, but our go--fill-forward-paragraph was giving up if
it didn't see a comment. Now we just fall back to forward-paragraph so
we can be sure we will make it out of the region.