update default go-packages-function to use go list
The previous default, go-packages-native doesn't work well with the Go
build cache and Go modules. Fixing it would require us to reimplement
a lot of the `go list` logic in elisp (and continue to maintain it for
future changes to the go tool). It doesn't make sense to do that as
the original reason for the elisp implementation (speed) is much less
of an issue when using modules.

Fixes #298