Run all tests in GH CI
And fix go-fill-paragraph-test.el to pass from the command line. Not
sure exactly what is wrong, but fill-paragraph wasn't doing the right
thing with a region selected when invoked from the command line. There
could be a real bug here, but I wasn't able to reproduce the failures
functionally in Emacs.

Closes: #308 [via git-merge-pr]