Fontify variable names.
We now fontify identifiers that declare variables with the
"font-lock-variable-name-face". That includes variables in "var"
decls, ":=" assignments, and function signatures.

I did a significant rewrite of function signature font locking. We now
use the built in "anchor" mechanism rather than our own anchored
searching. This simplified things a lot and means we no longer have a
limit to how many params we can fontify. Basically, there is a parent
matcher that looks for "(" that starts param lists. The child matcher
iterates over each param in the list. The key is that the child
matcher is able to keep state, so we can remember whether the param
list has names or not.

Fixes #21.