Fix multiline func signature fontification.
Now that we use the proper anchor mechanism, we can just have the
"pre" form return the position of the closing param list ")", and set
font-lock-multiline, and basically that is it.

I thought I was going to have to add something to
font-lock-extend-region-functions, but that didn't seem necessary. If
there are further cases of missing fontification we can try adding the
region extending function for param lists.