Fix fontification performance issue
When looking for the start of a param list, don't match "(" in
comments/strings. This was causing us to do a lot of extra work when
the comment was at the file scope.

What I think was happening was we would expand our multiline keyword
scope to the end of the buffer looking for the corresponding closing
")". This caused us to font lock much bigger parts of the buffer where
normally you only font lock a line or two at a time.

I also killed some dead code I noticed in go--match-decl.

Closes: #322 [via git-merge-pr]