Speed up slow fontification after unclosed parens.
To handle multi line param lists we "expand" the fontification scope
to the closing paren. If there is no balanced closing paren then we
expand the scope to the end of the file, which fontifies too much
stuff and makes typing very slow.

For example, typing is slow at <>:

func (myReceive<>

... lots of lines in rest of file ...

Add a quick fix to check that the paren is closed by the end of the
file. If it isn't closed, then don't expand the region. There is more
work to do here (e.g. func literals inside giant functions will still
be slow), but this should resolve the worst slowness.

Fixes #335.