Fix fontification of multiplicand after index expr.
In "foo[1]*10" we were interpreting "[1]*10" as an array with element
type "*10". At first I was going to fix the identifier regex to
disallow leading numbers (so "10" couldn't be a type name), but that
is involved and doesn't fix the similar case of "foo[1]*foo[2]".
Instead fix the array type detection to disallow an identifier
character preceding the opening bracket.

Fixes #382.