Fontify type args in new()/make()
This uses a cheeky approach where we piggyback off existing type
matching. After using the existing matcher to fontify "foo" in
make(foo[bar]), we then check if the next char is "[", and if so we
know it is a type list, so we fontify it as such.

This approach will be used in later commits for all the other simple
type matchers.