go-beginning-of-defun: only move to end of line when we're on a declaration
Some code (e.g. evil-mode) doesn't expect beginning-of-defun (with
positive arguments) to move the point forward. We move it to the end of
line to ensure we find the function declaration that point might already
be on. Do this conditionally to ensure we only move the point to eol
when a later re-search-backward will undo that motion.

Updates gh-186